==> Drug_Repositioning/8_new_training/1_binary_matrix.ipynb <==
import csv
import numpy as np
import os
import pandas as pd
import requests
import time

############# Asssociation ##############

asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/7_gold_association.tsv")

print len(asso)
asso.head(1)

disease_ids = asso['disease_id'].drop_duplicates()
dg_ids = asso['drug_id'].drop_duplicates()

print len(disease_ids), len(dg_ids)

binary_asso = pd.DataFrame(columns = disease_ids.tolist(), index = dg_ids.tolist())
binary_asso = binary_asso.fillna(0)

f = lambda x: 1 if x else 0
    
for di_id in disease_ids:
    di_df = asso[asso.disease_id == di_id]
    dg_id = di_df['drug_id']
    flags = dg_ids.isin(dg_id)
    oz_flags = map(f, flags)
    binary_asso[di_id] = oz_flags
                   

binary_asso.to_csv("/home/share/aschoi/data/Drug_Repositioning/7_new_association/8_gold_standard_association_bianry_matrix.tsv", sep='\t')
binary_asso.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/8_gold_standard_association_bianry_matrix.tsv", sep='\t')

########### drug ############

drug = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/7_2_gold_drug_feature_id.tsv")

print len(drug)
drug.head(1)

dg_ids = drug['DrugBank_ID'].drop_duplicates()
umls_ids = drug['umls_id'].drop_duplicates()

binary_drug = pd.DataFrame(columns = dg_ids.tolist(), index = umls_ids.tolist())
binary_drug = binary_drug.fillna(0)

f = lambda x: 1 if x else 0
    
for dg_id in dg_ids:
    dg_df = drug[drug.DrugBank_ID == dg_id]
    umls_id = dg_df['umls_id']
    flags = umls_ids.isin(umls_id)
    oz_flags = map(f, flags)
    binary_drug[dg_id] = oz_flags

binary_drug.to_csv("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/8_gold_standard_drug_binary_matrix.tsv", sep='\t')
binary_drug.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/8_gold_standard_drug_binary_matrix.tsv", sep='\t')

########### disease ############

disease = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/8_gold_disease_feature.tsv")

print len(disease)
disease.head(1)

disease_ids = disease['OMIM_ID'].drop_duplicates()
gene_ids = disease['DisGeNet_geneId'].drop_duplicates()
print len(disease_ids)
print len(gene_ids)

binary_disease = pd.DataFrame(columns = disease_ids.tolist(), index = gene_ids.tolist())
binary_disease = binary_disease.fillna(0)

f = lambda x: 1 if x else 0
    
for disease_id in disease_ids:
    disease_df = disease[disease.OMIM_ID == disease_id]
    gene_id = disease_df['DisGeNet_geneId']
    flags = gene_ids.isin(gene_id)
    oz_flags = map(f, flags)
    binary_disease[disease_id] = oz_flags

혹은,

omim_ids = disease['OMIM_ID']
gene_ids= disease['DisGeNet_geneId']

result = pd.DataFrame(columns = omim_ids.drop_duplicates().tolist(), index = gene_ids.drop_duplicates().tolist())
result = result.fillna(0)

for omim_id in omim_ids.drop_duplicates():
    found = disease[disease.OMIM_ID == omim_id]
    if found.empty:
        continue
    else:
        for index, row in found.iterrows():
            result[omim_id].loc[row['DisGeNet_geneId']] = 1

#실제 training에는 두번째 방법으로 만든 binary matrix를 쓴다.

result.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/8_gold_standard_disease_binary_matrix.tsv",sep ='\t')
result.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/8_gold_standard_disease_binary_matrix.tsv",sep ='\t')
