==> Drug_Repositioning/8_new_training/1_gold_descriptor.ipynb <==
import numpy as np
#import matplotlib.pyplot as plt

from sklearn.linear_model import LogisticRegression
from sklearn.preprocessing import StandardScaler
from sklearn import datasets
from sklearn import metrics
import pandas as pd
import time

association_data = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/8_gold_standard_association_bianry_matrix.tsv", index_col = 0)
drug = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/8_gold_standard_drug_binary_matrix.tsv", index_col = 0)
disease = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/8_gold_standard_disease_binary_matrix.tsv", index_col = 0)
df_gold_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/9_gold_learning_association.tsv")
df_indep_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/9_2_indep_learning_association.tsv")

print len(df_gold_asso)
df_gold_asso.head(1)

drug.head(1)

print df_gold_asso.shape
df_gold_asso.values

drug.index

print cnt

############### 1.13. 추가 ######### descriptor id도 같이
start = 0
end = 10001
cnt = 0
arr = np.empty((0,5797), dtype=np.int64)
arr = np.hstack((['drug_id', 'disease_id', 'association', drug.index, disease.index]))
for i in df_gold_asso.values[start:end]:
    drug_id = i[0]
    disease_id = i[1]
    association = i[2]
    feature = np.hstack([drug_id, disease_id, association, drug[drug_id], disease[str(disease_id)]])
    arr = np.vstack([arr, feature])
    cnt = cnt + 1
    if cnt % 1000 == 0:
        print "%d : " % cnt + time.strftime('%a %H:%M:%S')
print "end : %d : " % cnt + time.strftime('%a %H:%M:%S')
print cnt
print arr.shape

df_arr = pd.DataFrame(data = arr[1:],columns = arr[0])
df_arr.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_0_10000.tsv", sep='\t', index = False)

arr1 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_0_10000.tsv")
arr2 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_10001_20000.tsv")
arr3 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_20001_30000.tsv")
arr4 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_30001_40000.tsv")
arr5 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_40001_50000.tsv")
arr6 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_50001_60000.tsv")
arr7 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_60001_70000.tsv")
arr8 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_70001_80000.tsv")
arr9 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_80001_90000.tsv")
arr10 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_90001_100000.tsv")
arr11 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_100001_110685.tsv")

frames = [arr1, arr2, arr3, arr4, arr5, arr6, arr7, arr8, arr9, arr10, arr11]
df_arr_all = pd.concat(frames , axis=0)
print len(df_arr_all)
df_arr_all.head(1)

df_arr_all.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_combined.tsv", sep = '\t', index = False)

# 이전 버전
start = 0
end = 10001
cnt = 0
arr = np.empty((0,5797), dtype=np.int64)
print "start %d : " % cnt + time.strftime('%a %H:%M:%S')
for i in df_asso.values[start:end]:
    drug_id = i[0]
    disease_id = i[1]
    feature = np.hstack([drug[drug_id], disease[str(disease_id)]])
    arr = np.vstack([arr, feature])
    cnt = cnt + 1
    if cnt % 1000 == 0:
        print "%d : " % cnt + time.strftime('%a %H:%M:%S')
print "end : %d : " % cnt + time.strftime('%a %H:%M:%S')
print cnt
print arr.shape

# 이전 버전
print "start %d : " % cnt + time.strftime('%a %H:%M:%S')
np.savetxt("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_0_10000.tsv", arr, delimiter="\t")
print "end : %d : " % cnt + time.strftime('%a %H:%M:%S')

arr2 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_10001_20000.tsv", header = None)
arr3 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_20001_30000.tsv", header = None)
arr4 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_30001_40000.tsv", header = None)
arr5 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_40001_50000.tsv", header = None)
arr6 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_50001_60000.tsv", header = None)
arr7 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_60001_70000.tsv", header = None)
arr8 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_70001_80000.tsv", header = None)
arr9 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_80001_90000.tsv", header = None)
arr10 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_90001_100000.tsv", header = None)
arr11 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_100001_110685.tsv", header = None)

print len(arr), len(arr2), len(arr3), len(arr4), len(arr5), len(arr6), len(arr7), len(arr8), len(arr9), len(arr10), len(arr11)

frames = [arr2, arr3, arr4, arr5, arr6, arr7, arr8, arr9, arr10, arr11]
arr2_11 = pd.concat(frames)
print len(arr2_11)

features = np.vstack([arr,arr2_11.values])
features.shape

np.savetxt("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_combined.tsv", features, delimiter="\t")

y = df_asso.values[:, 2]
from sklearn.cross_validation import train_test_split
print "start : " + time.strftime('%a %H:%M:%S')
X_train, X_test, y_train, y_test = train_test_split(features, y, test_size = 0.25)
print "end : " + time.strftime('%a %H:%M:%S')

print X_train.shape, y_train.shape
print X_test.shape, y_test.shape
Y_train = list(y_train)
Y_test = list(y_test)

print sum(Y_train), sum(Y_test)

print "start : " + time.strftime('%a %H:%M:%S')
for i, C in enumerate((10000, 1000, 100, 10, 1, 0.1, 0.01, 0.001, 0.0001)):
    clf_l1_LR = LogisticRegression(C=C, penalty='l1', tol= 0.01)
    #clf_l1_LR.fit(X_train, y_train)
    clf_l1_LR.fit(X_train, Y_train)
    
    coef_l1_LR = clf_l1_LR.coef_.ravel()
    
    sparsity_l1_LR = np.mean(coef_l1_LR == 0) * 100

    #print "%.4lf : " % C + time.strftime('%a %H:%M:%S')
    
    print("C=%.4f" % C)
    print("Sparsity with L1 penalty: %.2f%%" % sparsity_l1_LR)
    print("score with L1 penalty: %.4f" % clf_l1_LR.score(X_train, Y_train))
    
    
    # 성능 측정-정확도 : training data에서의 정확도 츨정. classifier가 정확하게 분류한 인스턴스의 비율을 측정한다. --> 이걸로 측정하면 안됨.
    y_train_pred = clf_l1_LR.predict(X_train)
    print metrics.accuracy_score(Y_train, y_train_pred)
    
    # test data에서의 정확도 측정 : 일반적으로 training data에서보다 낮은 정확도를 보인다.
    y_pred = clf_l1_LR.predict(X_test)
    print metrics.accuracy_score(Y_test, y_pred)
    
    #평가함수
    print metrics.classification_report(Y_test, y_pred)
    
    
print "end : " + time.strftime('%a %H:%M:%S')
