==> Drug_Repositioning/8_new_training/1_herb_descriptor.ipynb <==
import numpy as np
import pandas as pd
import time

herb = pd.read_table("/DAS_Storage1/aschoi/data/COCONUT_Herbal/6_Herb/7_herb_compound_combined_common_phenotypic_effects.tsv")
herb_asso = pd.read_table("/DAS_Storage1/aschoi/data/COCONUT_Herbal/6_Herb/7_2_learning_herb_compound_disease_association.tsv")
disease = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/8_gold_disease_feature.tsv")
disgenet = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/curated_gene_disease_associations.tsv", skiprows = 21)

print len(herb)
print len(herb['compID'].drop_duplicates())
print len(herb['umls_id'].drop_duplicates())
herb.head(1)

disease.head(1)

print len(disease)
print len(disease['UMLS_concept_ID'].drop_duplicates())
print len(disease['OMIM_ID'].drop_duplicates())
print len(disease['DisGeNet_geneId'].drop_duplicates())
disease.head(1)

print len(disgenet)
print len(disgenet['diseaseId'].drop_duplicates())
print len(disgenet['geneId'].drop_duplicates())
disgenet['diseaseId']  = disgenet['diseaseId'].str.slice(start = 5)
disgenet.head(1)

print len(herb_asso)
print len(herb_asso['compID'].drop_duplicates()) # disease association이 존재하는 herbal compound(이 compound들은 drug feature처리가 되어있음.)
print len(herb_asso['umlsID'].drop_duplicates()) # disease의 수
herb_asso.head(1)

# herb_asso의 disease 중, feature를 가지고 있는 애들
k = herb_asso[herb_asso['umlsID'].isin(disgenet['diseaseId'])]
print len(k)
print len(k['compID'].drop_duplicates())
print len(k['umlsID'].drop_duplicates())

# disgenet 중에, herb_asso의 disease를 보기
c = disgenet[disgenet['diseaseId'].isin(herb_asso['umlsID'])]
print len(c['diseaseId'].drop_duplicates())
print len(c['geneId'].drop_duplicates())

# 둘이 join 함
combined = pd.merge(left = k, right = c, left_on = 'umlsID', right_on='diseaseId')
print len(combined)
print len(combined.drop_duplicates())
print len(combined.drop_duplicates(subset = ['compID', 'umlsID']))
print len(combined['compID'].drop_duplicates())
print len(combined['umlsID'].drop_duplicates())
print len(combined['diseaseId'].drop_duplicates())
print len(combined['geneId'].drop_duplicates())
combined.head(1)

# learning에 쓰이는 gold standard disease feature로 filter해야함.
h = combined[combined['geneId'].isin(disease['DisGeNet_geneId'])]
print len(h)
print len(h.drop_duplicates())
print len(h.drop_duplicates(subset = ['compID', 'umlsID']))
print len(h['compID'].drop_duplicates())
print len(h['umlsID'].drop_duplicates())
print len(h['geneId'].drop_duplicates())
h.head(1)

h.columns.values

filtered = h[['compID', 'effectType', 'evidence', 'phenID', 'reference', 'umlsID','diseaseId', 'geneId', 'score', 'geneName', 'description','diseaseName', 'sourceId']]
results = filtered[~filtered['umlsID'].isin(disease['UMLS_concept_ID'])]
print len(results)
print len(results.drop_duplicates())
print len(results.drop_duplicates(subset = ['compID', 'umlsID']))
print len(results.drop_duplicates(subset = 'compID'))
print len(results.drop_duplicates(subset = 'umlsID'))
print len(results.drop_duplicates(subset = 'geneId'))

results.to_csv("/home/share/aschoi/nas/users/asolchoi/data/COCONUT_Herbal/6_Herb/7_3_learning_herb_compound_disease_association_final.tsv", index=False)
results.to_csv("/DAS_Storage1/aschoi/data/COCONUT_Herbal/6_Herb/7_3_learning_herb_compound_disease_association_final.tsv", index=False)

# learning에 쓰일 수 있는 형태로 만들기
herb_feature = pd.read_table("/DAS_Storage1/aschoi/data/COCONUT_Herbal/6_Herb/7_herb_compound_combined_common_phenotypic_effects.tsv")
common_feature = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/7_2_gold_drug_feature_id.tsv")

disease_mat = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/8_gold_standard_disease_binary_matrix.tsv", index_col=0)
drug_mat = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/8_gold_standard_drug_binary_matrix.tsv", index_col = 0)

disease_mat.head(1)

drug_feature = drug_mat.index
disease_feature = disease_mat.index
print len(drug_feature), len(disease_feature)

herb_disease_asso = results[['compID', 'umlsID']].drop_duplicates()
herb_disease_asso.index = range(len(herb_disease_asso))

print len(herb_disease_asso)
print len(herb_disease_asso.drop_duplicates(subset = 'compID'))
print len(herb_disease_asso.drop_duplicates(subset = 'umlsID'))
herb_disease_asso.head(1)


print len(herb_feature)
print len(herb_feature.drop_duplicates(subset = 'compID'))
print len(herb_feature.drop_duplicates(subset = 'umls_id'))
herb_feature.head(1)

f = lambda x: 1 if x else 0

cnt = 0
#arr = np.empty((0,5797), dtype=np.int64)
arr = np.hstack(('herb_compound_id', 'disease_id', drug_mat.index, disease_mat.index))
for index, row in herb_disease_asso.iterrows():
    compID = row['compID']
    umlsID = row['umlsID']
    
    temp_herb_compound = herb_feature[herb_feature.compID == compID]['umls_id']
    hc_flags = drug_mat.index.isin(temp_herb_compound)
    herb_compound_flags = map(f, hc_flags)
    
    temp_herb_disease = results[results.umlsID == umlsID]['geneId'].drop_duplicates()
    hd_flags = disease_mat.index.isin(temp_herb_disease)
    herb_disease_flags = map(f, hd_flags)
    
    feature = np.hstack([compID, umlsID, np.array(herb_compound_flags), np.array(herb_disease_flags)])
    arr = np.vstack([arr, feature])
    cnt = cnt + 1
    if cnt % 1000 == 0 :
        print "%d : " % cnt + time.strftime('%a %H:%M:%S')
print "end : %d : " % cnt + time.strftime('%a %H:%M:%S')
print cnt
print arr.shape

df_arr = pd.DataFrame(data = arr[1:], columns = arr[0])
print len(df_arr)
df_arr.head(1)

df_arr.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/3_herb_asso_descriptor.tsv", sep='\t', index=False)
df_arr.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/8_new_training/3_herb_asso_descriptor.tsv", sep='\t', index=False)

#이전버전

herb_compound_descriptor = pd.concat([herb_disease_asso, df_arr], axis=1)

herb_compound_descriptor.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/3_herb_asso_descriptor.tsv", sep='\t', index=False)
herb_compound_descriptor.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/8_new_training/3_herb_asso_descriptor.tsv", sep='\t', index=False)

print len(herb_compound_descriptor)
herb_compound_descriptor.head(1)
