==> Drug_Repositioning/8_new_training/1_indep_descriptor.ipynb <==
import numpy as np
import pandas as pd
from sklearn.linear_model import LogisticRegression
from sklearn.preprocessing import StandardScaler
import time

asso = pd.read_excel("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/MBiRW-master/Datasets/Datasets_indep/DiDrMat.xlsx", header=None)
#asso = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/MBiRW-master/Datasets/Datasets_indep/DiDrMat.txt")
drug = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/MBiRW-master/Datasets/Datasets_indep/DrugsName", header=None)
disease = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/MBiRW-master/Datasets/Datasets_indep/DiseasesName", header=None)

drug = drug.rename(columns = {0 : 'drug_id'})
disease = disease.rename(columns = {0 : 'disease_id'})

asso.columns = drug['drug_id'].tolist()

print len(disease)
print len(drug)

asso.index = disease['disease_id'].tolist()

df_asso = pd.DataFrame()
for disease_id in asso.index.values:
    for drug_id in asso.columns.values:
        temp_df = pd.DataFrame([[drug_id, disease_id, asso[drug_id].loc[disease_id]]], columns = ['drug', 'disease', 'association'])
        df_asso = df_asso.append(temp_df)

print len(df_asso[df_asso.association == 1]) # positive
print len(df_asso[df_asso.association == 0]) # negative

df_asso.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/MBiRW-master/Datasets/Datasets_indep/indep_asso.tsv", sep='\t', index=False)
df_asso.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/MBiRW-master/Datasets/Datasets_indep/indep_asso.tsv", sep='\t', index=False)

# 사용할 수 있는 애들

sider_DrugBank_id = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/3_3_SIDER_STITCHid_DrugBankid.tsv")
offside_DrugBank_id = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/4_3_OFFSIDE_STITCHid_DrugBankid.tsv")
sider_feature = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/3_4_SIDER_db_side_effect.tsv")
offside_feature = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/4_4_OFFSIDE_db_side_effect.tsv")

sider_found = sider_DrugBank_id[sider_DrugBank_id['DrugBank_ID'].isin(drug['drug_id'])]
print len(sider_found)
print len(sider_found.drop_duplicates(subset = 'DrugBank_ID'))
sider_found.head(1)

offside_found = offside_DrugBank_id[offside_DrugBank_id['DrugBank_ID'].isin(drug['drug_id'])]
print len(offside_found)
print len(offside_found.drop_duplicates(subset = 'DrugBank_ID'))
offside_found.head(1)

offside_feature.head(1)

sider_result = pd.DataFrame()
for index, row in sider_found.iterrows():
    li_row = row.tolist()
    flat = li_row[0]
    stereo = li_row[1]
    drugbank_id = li_row[2]
    filtered = sider_feature[(sider_feature.STITCH_compound_id_flat == flat) & (sider_feature.STITCH_compound_id_stereo == stereo)]
    for index2, row2 in filtered.iterrows():
        umls_id = row2['UMLS_concept_id_for_MedDRA']
        temp_df = pd.DataFrame([[drugbank_id, umls_id]], columns=['DrugBank_ID', 'umls_id'])
        sider_result = sider_result.append(temp_df)

print len(sider_result.drop_duplicates(subset = 'DrugBank_ID'))
print len(sider_result.drop_duplicates(subset = 'umls_id'))

offside_result = pd.DataFrame()
for index, row in offside_found.iterrows():
    li_row = row.tolist()
    stereo = li_row[0]
    drugbank_id = li_row[1]
    filtered = offside_feature[offside_feature.stitch_id == stereo]
    for index2, row2 in filtered.iterrows():
        umls_id = row2['umls_id']
        temp_df = pd.DataFrame([[drugbank_id, umls_id]], columns=['DrugBank_ID', 'umls_id'])
        offside_result = offside_result.append(temp_df)

print len(offside_result.drop_duplicates(subset = 'DrugBank_ID'))
print len(offside_result.drop_duplicates(subset = 'umls_id'))

frames = [sider_result, offside_result]
combined = pd.concat(frames).sort_values(by = 'DrugBank_ID').drop_duplicates()

print len(combined)
print len(combined.drop_duplicates(subset = 'DrugBank_ID')) # drug의 수

combined.head(1)

gold_drug_feature = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/7_2_gold_drug_feature_id.tsv")

print len(gold_drug_feature)
print len(gold_drug_feature.drop_duplicates(subset = 'umls_id'))
gold_drug_feature.head(1)

common_feature = gold_drug_feature['umls_id'].drop_duplicates()

filtered = combined[combined['umls_id'].isin(common_feature)]

print len(filtered)
print len(filtered.drop_duplicates(subset = 'DrugBank_ID')) # indep의 drug 수 
print len(filtered.drop_duplicates(subset = 'umls_id')) # 공통되는 feature의 수
filtered.head(1)

final_asso = df_asso[df_asso['drug'].isin(filtered['DrugBank_ID'])]
print len(final_asso[final_asso.association == 0]) # negative
print len(final_asso[final_asso.association == 1]) # positive
print len(final_asso.drop_duplicates('drug'))
print len(final_asso.drop_duplicates('disease'))
final_asso.head(1)

di_mimNum = final_asso['disease'].str[1:]
int_di_mimNum = list()
for i in di_mimNum:
    int_di_mimNum.append (int(i))
print len(di_mimNum)
print len(int_di_mimNum)

final_asso['disease'] = int_di_mimNum

여기서 positive만 고려한것이 필요함.

df_positive = final_asso[final_asso.association == 1]
print len(df_positive['drug'].drop_duplicates()) #drug의 수 
print len(df_positive['disease'].drop_duplicates()) #disease의 수
print len(df_positive) # positive association의 수 
df_positive = df_positive.rename(columns={'drug' : 'drug_id', 'disease' : 'disease_id'})
df_positive.head(1)

#이 중에 disease feature가 없는 disease가 있음. 걸러내야함.
disease = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/8_gold_standard_disease_binary_matrix.tsv", index_col=0)

print len(disease.columns.values) # gold standard disease의 수
disease_entry = pd.Series(disease.columns.values)

print len(df_positive['disease_id'].drop_duplicates()) # 현재 indep disease의 수 
indep_disease_entry = df_positive['disease_id'].drop_duplicates()
indep_feature_disease_entry = disease_entry[disease_entry.isin(indep_disease_entry.values.astype(str))] # gold standard disease에서 indep disease가 포함되는 것만 거른것
# --> gold standard에서 거른것은, indep disease는 기존의 disease와 같기 때문. 그리고 gold standard disease는 disease feature가 있는 것이 확실한 애들이기 때문.

final_indep_asso = df_positive[df_positive['disease_id'].isin(indep_feature_disease_entry.values.astype(int))]

print len(final_indep_asso)
print len(final_indep_asso[final_indep_asso.association == 0])
print len(final_indep_asso[final_indep_asso.association == 1])
print len(final_indep_asso.drop_duplicates(subset = 'drug_id'))
print len(final_indep_asso.drop_duplicates(subset = 'disease_id'))
final_indep_asso.head(1)

final_indep_asso[['drug_id', 'disease_id']].to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/7_2_indep_association.tsv", sep='\t', index=False)
final_indep_asso[['drug_id', 'disease_id']].to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/7_2_indep_ssociation.tsv", sep='\t', index=False)

#learning 꼴로 만들기.
learning_asso = final_asso[final_asso['disease'].isin(final_indep_asso['disease_id'])] # final_asso에서 disease id filter
learning_asso = learning_asso[learning_asso['drug'].isin(final_indep_asso['drug_id'])] # final_asso에서 drug id filter
print len(learning_asso.drop_duplicates(subset = 'drug'))
print len(learning_asso.drop_duplicates(subset = 'disease'))
print len(learning_asso[learning_asso.association == 1])
learning_asso.head(1)

learning_asso.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/9_2_indep_learning_association.tsv", sep='\t', index=False)
learning_asso.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/9_2_indep_learning_association.tsv", sep='\t', index=False)

#binary matrix 꼴로 만들기 (drug, association)

filtered.index = range(len(filtered))
filtered.head(1)

learning_asso.index = range(len(learning_asso))
learning_asso.head(1)

df_learnig_drug_feature = filtered[filtered['DrugBank_ID'].isin(learning_asso['drug'])]
print len(df_learnig_drug_feature)
print len(df_learnig_drug_feature.drop_duplicates(subset ='DrugBank_ID'))
print len(df_learnig_drug_feature.drop_duplicates(subset ='umls_id'))
df_learnig_drug_feature.head(1)

df_learnig_drug_feature.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/7_3_indep_drug_feature_id.tsv", sep = '\t', index = False)
df_learnig_drug_feature.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/7_3_indep_drug_feature_id.tsv", sep = '\t', index = False)

#drug
#dg_ids = filtered['DrugBank_ID'].drop_duplicates()
dg_ids = df_learnig_drug_feature['DrugBank_ID'].drop_duplicates()
umls_ids = df_learnig_drug_feature['umls_id'].drop_duplicates()

binary_drug = pd.DataFrame(columns = dg_ids.tolist(), index = umls_ids.tolist())
binary_drug = binary_drug.fillna(0)

f = lambda x: 1 if x else 0
for dg_id in dg_ids:
    dg_df = filtered[filtered.DrugBank_ID == dg_id]
    umls_id = dg_df['umls_id']
    flags = umls_ids.isin(umls_id)
    oz_flags = map(f, flags)
    binary_drug[dg_id] = oz_flags
print binary_drug.shape

binary_drug.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/8_2_indep_standard_drug_binary_matrix.tsv", sep='\t')
binary_drug.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/8_2_indep_standard_drug_binary_matrix.tsv", sep='\t')

#association

disease_ids = learning_asso['disease'].drop_duplicates()
dg_ids = learning_asso['drug'].drop_duplicates()

binary_asso = pd.DataFrame(columns = disease_ids.tolist(), index = dg_ids.tolist())
binary_asso = binary_asso.fillna(0)

f = lambda x: 1 if x else 0
    
for di_id in disease_ids:
    di_df = learning_asso[learning_asso.disease == di_id]
    dg_id = di_df['drug']
    flags = dg_ids.isin(dg_id)
    oz_flags = map(f, flags)
    binary_asso[di_id] = oz_flags 
print binary_asso.shape

binary_asso.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/8_2_indep_association_bianry_matrix.tsv", sep='\t')
binary_asso.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/8_2_indep_association_bianry_matrix.tsv", sep='\t')

#descriptor

indep_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/9_2_indep_learning_association.tsv")
indep_drug = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/8_2_indep_standard_drug_binary_matrix.tsv", index_col=0)
disease = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/8_gold_standard_disease_binary_matrix.tsv", index_col=0)
drug = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/8_gold_standard_drug_binary_matrix.tsv", index_col = 0)

print len(indep_asso)
print len(indep_asso[indep_asso.association == 1]) # positive
print len(indep_asso[indep_asso.association == 0]) # negative
print len(indep_asso['drug'].drop_duplicates())
print len(indep_asso['disease'].drop_duplicates())
print indep_asso.shape
indep_asso.head(1)

print indep_drug.shape
indep_drug.head(1)

cnt = 0
arr = np.empty((0,5797), dtype=np.int64)
arr = np.hstack(('drug_id', 'disease_id', 'association', drug.index, disease.index))
f = lambda x: 1 if x else 0
print "start %d : " % cnt + time.strftime('%a %H:%M:%S')
for i in indep_asso.values:
    drug_id = i[0]
    disease_id = i[1]
    association = i[2]
    
    temp_indep_drug = indep_drug[drug_id][indep_drug[drug_id] == 1] # drug_id에 해당하는 umls_id들만 추린것.(umls_id는 index임)
    flags = drug.index.isin(temp_indep_drug.index)
    indep_drug_flags = map(f, flags)
    feature = np.hstack([drug_id, disease_id, association, np.array(indep_drug_flags), disease[str(disease_id)]])
    arr = np.vstack([arr, feature])
    cnt = cnt + 1
    if cnt % 1000 == 0:
        print "%d : " % cnt + time.strftime('%a %H:%M:%S')
print "end : %d : " % cnt + time.strftime('%a %H:%M:%S')
print cnt
print arr.shape

df_arr = pd.DataFrame(data = arr[1:], columns = arr[0])
print len(df_arr)
df_arr.head(1)

df_arr.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_indep_asso_descriptor.tsv", sep='\t', index=False)
df_arr.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/8_new_training/2_indep_asso_descriptor.tsv", sep='\t', index=False)

indep_asso_descriptor.head(1)
