==> Drug_Repositioning/8_new_training/2_negative_descriptor.ipynb <==
import numpy as np
#import matplotlib.pyplot as plt
import pandas as pd
import time

#########2017.1.17. positive에는 많고, negative feature에는 적은 feature를 선별해야 함.

gold = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_combined.tsv")
pos_feature_sum = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/3_2_pos_feature_sum.tsv", header=None)
neg_feature_sum = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/3_1_neg_feature_sum.tsv", header=None)

gold_positive = gold[gold.association == 1]
gold_negative = gold[gold.association == 0]
print len(gold_positive)
print len(gold_negative)

# pos feature가 negative feature 순위가 어떻게 되는지 확인함. 값이 작을 수록 neg에서도 자주 나오는 feature임.

neg_index = list()
for i in pos_feature_sum[0]:
    idx = neg_feature_sum[0].tolist().index(i)
    neg_index.append(idx)

# neg_feature_sum을 pos_feature_sum의 feature 순위대로 정렬한것

#temp = pd.DataFrame({'feature':list(neg_feature_sum[0].values), 'sum':list(neg_feature_sum[1].values)})
#temp.feature = temp.feature.astype("category")
#temp.feature.cat.set_categories(list(pos_feature_sum[0].values), inplace=True)
#temp = temp.sort_values(["feature"])
#temp.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/8_new_training/3_3_neg_sum_by_pos.tsv", sep="\t", index=False)

#위 결과를 파일로 저장했음. 그 파일을 읽어오기
neg_sum_by_pos = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/3_3__neg_sum_by_pos.tsv")

# percentage를 적용할 df 만듬. feature 순위, feature sum, neg-pos feature sum 정보가 있음. absolute values는 rank 뺀걸로 만듬.

#positive_rank= range(5798)[1:]
#negative_rank = [x+1 for x in neg_index]
#abs_difference = abs(pd.Series(positive_rank) - pd.Series(negative_rank))
#df = pd.DataFrame(data = {'feature' : list(pos_feature_sum[0].values), 
#              'positive_sum' : list(pos_feature_sum[1].values), 'positive_cover' : (pos_feature_sum[1]/1434*100).values, 'positive_rank' : positive_rank,
#              'negative_sum' : neg_sum_by_pos['sum'].values, 'negative_cover' : (neg_sum_by_pos['sum']/109251*100).values, 'negative_rank' : negative_rank,
#              'absolute_difference' : abs_difference.values}, index = range(5797))
#df = df[['feature', 'positive_sum', 'positive_cover', 'negative_sum', 'negative_cover', 'positive_rank', 'negative_rank', 'absolute_difference']]

#df.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/8_new_training/3_4_neg_pos_feature_rank.tsv", sep='\t', index=False)
#df.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/3_4_neg_pos_feature_rank.tsv", sep='\t', index=False)

#위 결과를 파일로 저장했음. 그 파일을 읽어오기
df = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/3_4_neg_pos_feature_rank.tsv")

sorted_df = df.sort_values(by = 'absolute_difference', ascending = False)
sorted_df.head(1)

over_cutoff = sorted_df[:int(round(len(sorted_df)*0.01))]
print len(over_cutoff)
print len(over_cutoff[over_cutoff.positive_rank < over_cutoff.negative_rank])
over_cutoff[over_cutoff.positive_rank < over_cutoff.negative_rank].sort_values(by = 'positive_cover', ascending = False).head(5)

print len(neg_sum_by_pos)
neg_sum_by_pos.head(1)

# percentage를 적용할 df 만듬. feature 순위, feature sum, neg-pos feature sum 정보가 있음. absolute values는  ocver 범위 뺀걸로 만듬.

#positive_rank= range(5798)[1:]
#negative_rank = [x+1 for x in neg_index]
#positive_cover = (pos_feature_sum[1]/1434*100).values
#negative_cover = (neg_sum_by_pos['sum']/109251*100).values
#abs_difference2 = abs(pd.Series(positive_cover) - pd.Series(negative_cover))
#df2 = pd.DataFrame(data = {'feature' : list(pos_feature_sum[0].values), 
#              'positive_sum' : list(pos_feature_sum[1].values), 'positive_cover' : positive_cover, 'positive_rank' : positive_rank,
#              'negative_sum' : neg_sum_by_pos['sum'].values, 'negative_cover' : negative_cover, 'negative_rank' : negative_rank,
#              'absolute_difference' : abs_difference2.values}, index = range(5797))
#df2 = df2[['feature', 'positive_sum', 'positive_cover', 'negative_sum', 'negative_cover', 'positive_rank', 'negative_rank', 'absolute_difference']]

#df2.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/8_new_training/3_5_neg_pos_feature_percent.tsv", sep='\t', index=False)
#df2.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/3_5_neg_pos_feature_percent.tsv", sep='\t', index=False)

#위 결과를 파일로 저장했음. 그 파일을 읽어오기
df2 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/3_5_neg_pos_feature_percent.tsv")

sorted_df2 = df2.sort_values(by = 'absolute_difference', ascending = False)
print len(sorted_df)
sorted_df2.head(1)

over_cutoff2 = sorted_df2[:int(round(len(sorted_df)*0.001))]
print len(over_cutoff2)
# a나 b나 같은 결과임.
#a = over_cutoff2[over_cutoff2.positive_rank < over_cutoff2.negative_rank].sort_values(by = 'positive_cover', ascending = False)
b = over_cutoff2[over_cutoff2.positive_cover > over_cutoff2.negative_cover].sort_values(by = 'positive_cover', ascending = False)
print len(b)
b.head(10)

positive_feature = b['feature']
positive_feature.index = range(len(positive_feature))
for i in range(len(positive_feature)):
    feature_id = positive_feature[i]
    if i == 0 :
        temp = gold_negative[gold_negative[feature_id] == 0]
    else :
        temp = temp[temp[feature_id] == 0]
#길이 제한 없음 최종 temp가 0이면 0되기 전에 끝내야함.

print len(gold_negative)
print len(temp)
temp.head(1)

#combine positive, negative
temp.index = range(len(temp))
frames = [gold_positive, temp]
combined = pd.concat(frames, axis = 0)
combined.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/4_gold_descriptor_revised_0.1%.tsv", sep='\t', index=False)
combined.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/8_new_training/4_gold_descriptor_revised_0.1%.tsv", sep='\t', index=False)

# 2017.1.16. 잘못된 descriptor (pos_feature_sum, neg_feature_sum 구하는 부분

)

gold = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_1_descriptor_combined.tsv")
indep = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_indep_asso_descriptor.tsv")
herb = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/3_herb_asso_descriptor.tsv")

gold_positive = gold[gold.association == 1]
gold_negative = gold[gold.association == 0]
print len(gold_positive)
print len(gold_negative)


features = gold.columns.values[3:].tolist()
df_positive_feature = gold_positive[features]
df_negative_feature = gold_negative[features]
pos_feature_sum = df_positive_feature.sum(axis=0)
neg_feature_sum = df_negative_feature.sum(axis=0)

pos_feature_sum = pos_feature_sum.sort_values(ascending=False)
neg_feature_sum = neg_feature_sum.sort_values(ascending=False)

positive_length = len(gold_positive)
percentage = 70
cutoff = positive_length * percentage / 100
filtered = pos_feature_sum[pos_feature_sum>cutoff]
print cutoff
print len(filtered)

for i in range(len(filtered.index)):
    feature_id = filtered.index[i]
    if i == 0 :
        temp = gold_negative[gold_negative[feature_id] == 0]
    else :
        temp = temp[temp[feature_id] == 0]
#길이 제한 없음 최종 temp가 0이면 0되기 전에 끝내야함.

print len(temp)

#combine positive, negative #이거 잘못된거임.
temp.index = range(len(temp))
frames = [gold_positive, temp]
combined = pd.concat(frames, axis = 0)
combined.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/4_gold_descriptor_revised.tsv", sep='\t', index=False)
combined.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/8_new_training/4_gold_descriptor_revised.tsv", sep='\t', index=False)
