==> Drug_Repositioning/8_new_training/4_PREDICT_descriptor.ipynb <==
import numpy as np
import os
import pandas as pd
import time
from itertools import product
import math

#### disease similarities 합치는 과정
#disease_ph = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_gold_disease_similarity_mimminerc.tsv", index_col=0)
#disease_ge = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_gold_disease_similarity_genetic.tsv")
#disease_ph2 = pd.DataFrame()
#for i in disease_ph.columns.values:
#    for j in disease_ph.index.values:
#        temp_df = pd.DataFrame(data = [[int(i), int(j), disease_ph[[i]].loc[j].values[0]]], columns = ['Disease_ID1', 'Disease_ID2', 'Similarity_Score'] )
#        disease_ph2 = disease_ph2.append(temp_df)
#df_disease_sim = pd.merge(left = disease_ph2, right= disease_ge, left_on = ['Disease_ID1', 'Disease_ID2'], right_on = ['Disease_ID1', 'Disease_ID2'])
#df_disease_sim.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/9_gold_disease_similarity_whole.tsv", sep='\t', index=False)
#df_disease_sim.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_gold_disease_similarity_whole.tsv", sep='\t', index=False)
#df_disease_sim = df_disease_sim.rename(columns = {'Similarity_Score_x':'PheSim', 'Similarity_Score_y':'GeSim'})
df_disease_sim = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_gold_disease_similarity_whole.tsv")

#### drug similarities 합치는 과정
#drug_ch = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_gold_drug_similarity_structure.tsv")
#drug_ph = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_gold_drug_similarity_phenotype.tsv")
#df_drug_sim = pd.merge(left = drug_ch, right= drug_ph, left_on = ['DrugBank_ID1', 'DrugBank_ID2'], right_on = ['DrugBank_ID1', 'DrugBank_ID2'])
#df_drug_sim = df_drug_sim.rename(columns = {'Similarity_Score_x' : 'CheSim', 'Similarity_Score_y':'PheSim'})
#df_drug_sim.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/9_gold_drug_similarity_whole.tsv", index=False, sep='\t')
#df_drug_sim.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_gold_drug_similarity_whole.tsv", index=False, sep='\t')
df_drug_sim = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_gold_drug_similarity_whole.tsv")

#### PREDICT용 새로운 association
#asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/9_gold_learning_association.tsv")
#new_asso = asso[asso['drug'].isin(df_drug_sim['DrugBank_ID1'].drop_duplicates())]
#new_asso.index = range(len(new_asso))
#new_asso.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/10_PREDICT_association.tsv", sep='\t', index=False)
#new_asso.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/10_PREDICT_association.tsv", sep='\t', index=False)
new_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/10_PREDICT_association.tsv")

print len(df_drug_sim)
df_drug_sim.head(1)

print len(df_disease_sim)
df_disease_sim.head(1)

print len(new_asso)
new_asso.head(1)

pos = new_asso[new_asso.association == 1]
neg = new_asso[new_asso.association == 0]
print len(pos)
print len(neg)

pos.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/5_PREDICT_descriptor/test_pos.tsv", sep='\t', index=False)
neg.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/8_new_training/test_neg.tsv", sep='\t', index=False)

new_asso_li = list()
for index, row in new_asso.iterrows():
    query_drug = row['drug']
    query_disease = row['disease']
    query_asso = row['association']
    new_asso_li.append((query_drug, query_disease, query_asso))

new_pos_li = list()
for index2, row2 in pos.iterrows():
    pos_drug = row2['drug']
    pos_disease = row2['disease']
    new_pos_li.append((pos_drug, pos_disease)) 

print 'start : ' + time.strftime('%a %H:%M:%S')
desc_pair = list(product(new_asso_li, new_pos_li))
print 'end : ' + time.strftime('%a %H:%M:%S')

print len(desc_pair)
desc_pair[1395:1399]

descriptor = pd.DataFrame()
cnt = 0
print 'start : ' + time.strftime('%a %H:%M:%S')
for pair in desc_pair:
    query_drug = pair[0][0]
    query_disease = pair[0][1]
    query_asso = pair[0][2]
    pos_drug = pair[1][0]
    pos_disease = pair[1][1]
    if (query_drug == pos_drug) & (query_disease == pos_disease):
            continue
    else:
        found_drug = df_drug_sim[(df_drug_sim.DrugBank_ID1 == query_drug) & (df_drug_sim.DrugBank_ID2 == pos_drug)]
        drug_cheSim = found_drug['CheSim'].values[0]
        drug_pheSim = found_drug['PheSim'].values[0]
    
        found_disease = df_disease_sim[(df_disease_sim.Disease_ID1 == query_disease) & (df_disease_sim.Disease_ID2 == pos_disease)]
        disease_pheSim = found_disease['PheSim'].values[0]
        disease_geSim = found_disease['GeSim'].values[0]
    
        drug_sim = [drug_cheSim, drug_pheSim]
        disease_sim = [disease_geSim, disease_pheSim]
        
        sim = [math.sqrt(x[0]*x[1]) for x in list(product(drug_sim, disease_sim))]
        temp_df = pd.DataFrame(data = [[query_drug, query_disease, query_asso, sim[0], sim[1], sim[2], sim[3]]],
                           columns=['drug', 'disease', 'association', 'DrChe_DiGe', 'DrChe_DiPhe', 'DrPhe_DiGe', 'DrPhe_DiPhe'])
        descriptor = descriptor.append(temp_df)
        cnt = cnt +1
        if cnt % 1000 == 0:
            print "{} : ".format(cnt) + time.strftime('%a %H:%M:%S')
print 'end : ' + time.strftime('%a %H:%M:%S')

cnt

descriptor.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/5_PREDICT_descriptor/pre_desc_0_129824.tsv", sep='\t', index=False)

descriptor.head()

print len(descriptor[(descriptor.drug == 'DB00190') & (descriptor.disease == 102300)]) # known
print len(descriptor[(descriptor.drug == 'DB00190') & (descriptor.disease == 102400)]) # unknwon
print len(descriptor[(descriptor.drug == 'DB00190') & (descriptor.disease == 173200)]) # unknwon 되다만거

real_desc = descriptor.groupby(['drug','disease']).max()
#real_desc.loc['DB00190',102300] # data 접근시
## 밑에거 잘라내기. partial slicing
real_desc = real_desc.loc[('DB00190', 102300):('DB00190', 171300)] # partial slicing (range로)

real_desc.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/5_PREDICT_descriptor/real_desc_0_129824.tsv", sep='\t')
real_desc.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/8_new_training/5_PREDICT_descriptor/real_desc_0_129824.tsv", sep='\t')

##### similarity tensor product 전까지만. 
descriptor = pd.DataFrame()
cnt = 0
print 'start : ' + time.strftime('%a %H:%M:%S')
bef_descriptor = list()
for index, row in new_asso.iterrows(): # all association
    print time.strftime('%a %H:%M:%S')
    query_drug = row['drug']
    query_disease = row['disease']
    query_asso = row['association']
    similarities = list()
    for index2, row2, in pos.iterrows(): # known association
        pos_drug = row2['drug']
        pos_disease = row2['disease']
        if (query_drug == pos_drug) & (query_disease == pos_disease):
            continue
        else:
            drug_cheSim = drug_ch[(drug_ch.DrugBank_ID1 == query_drug) & (drug_ch.DrugBank_ID2 == pos_drug)]['Similarity_Score']
            drug_pheSim = drug_ph[(drug_ph.DrugBank_ID1 == query_drug) & (drug_ph.DrugBank_ID2 == pos_drug)]['Similarity_Score']
            disease_pheSim = disease_ph[[str(query_disease)]].loc[pos_disease]
            disease_geSim = disease_ge[(disease_ge.Disease_ID1 == query_disease) & (disease_ge.Disease_ID2 == pos_disease)]['Similarity_Score']
            drug_sim = [drug_cheSim.values[0], drug_pheSim.values[0]]
            disease_sim = [disease_geSim.values[0], disease_pheSim.values[0]]
            similarities.append([drug_sim, disease_sim])
            #similarities.append([math.sqrt(x[0]*x[1]) for x in list(product(drug_sim, disease_sim))])
    bef_descriptor.append(similarities)    
    cnt = cnt + 1
    print time.strftime('%a %H:%M:%S')
    if cnt % 1000 == 0 :
        print "cnt : {} ".format(cnt) + time.strftime('%a %H:%M:%S')
print 'end : ' + time.strftime('%a %H:%M:%S')

print query_disease
print pos_disease
print disease_ph[[str(query_disease)]].loc[pos_disease]

##### similarity tensor product 후 score까지만. 
cnt = 0
print time.strftime('%a %H:%M:%S')
bef_descriptor = list()
for index, row in new_asso.iterrows(): # all association
    query_drug = row['drug']
    query_disease = row['disease']
    query_asso = row['association']
    similarities = list()
    print time.strftime('%a %H:%M:%S')
    for index2, row2, in pos.iterrows(): # known association
        pos_drug = row2['drug']
        pos_disease = row2['disease']
        if (query_drug == pos_drug) & (query_disease == pos_disease):
            continue
        else:
            drug_cheSim = drug_ch[(drug_ch.DrugBank_ID1 == query_drug) & (drug_ch.DrugBank_ID2 == pos_drug)]['Similarity_Score']
            drug_pheSim = drug_ph[(drug_ph.DrugBank_ID1 == query_drug) & (drug_ph.DrugBank_ID2 == pos_drug)]['Similarity_Score']
            disease_pheSim = disease_ph[[str(query_disease)]].loc[pos_disease]
            disease_geSim = disease_ge[(disease_ge.Disease_ID1 == query_disease) & (disease_ge.Disease_ID2 == pos_disease)]['Similarity_Score']
            drug_sim = [drug_cheSim.values[0], drug_pheSim.values[0]]
            disease_sim = [disease_geSim.values[0], disease_pheSim.values[0]]
            similarities.append([math.sqrt(x[0]*x[1]) for x in list(product(drug_sim, disease_sim))])

    bef_descriptor.append(similarities)
    print time.strftime('%a %H:%M:%S')
    break
    cnt = cnt + 1
    if cnt % 1000 == 0 :
        print "cnt : {} ".format(cnt) + time.strftime('%a %H:%M:%S')
print time.strftime('%a %H:%M:%S')

cnt
