==> Drug_Repositioning/8_new_training/5_SIDER_indication.ipynb <==
import numpy as np
import pandas as pd
import time

meddra = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/meddra_all_indications_header.tsv")
asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/7_gold_association.tsv")
omim_umls = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/1_omim_umls_mapping_association.tsv")
sider_drugbank = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/3_3_SIDER_STITCHid_DrugBankid.tsv")

print len(asso['drug_id'].drop_duplicates())
print len(asso['disease_id'].drop_duplicates())
asso.head(1)

omim_umls.head(1)

sider_drugbank.head(1)

#SIDER에서 제공하는 indication 정보가 training에도 쓰이는지 확인

#meddra = meddra[meddra.MedDRA_concept_type == 'PT'][['STITCH_compound_id_flat', 'UMLS_concept_id_for_MedDRA']]
print len(meddra), len(meddra['STITCH_compound_id_flat'].drop_duplicates())
meddra.head(1)

meddra_drugbank = sider_drugbank[sider_drugbank['STITCH_compound_id_flat'].isin(meddra['STITCH_compound_id_flat'])]
print len(meddra_drugbank)
print len(meddra_drugbank.drop_duplicates('STITCH_compound_id_flat'))
#meddra_drugbank[meddra_drugbank.duplicated(subset = 'STITCH_compound_id_flat', keep=False)].sort_values(by = 'STITCH_compound_id_flat')
meddra_drugbank.head(1)

len(asso[asso['drug_id'].isin(meddra_drugbank['DrugBank_ID'])].drop_duplicates('drug_id'))

disease_umls = omim_umls[omim_umls['OMIM ID'].isin(asso['disease_id'])]
print len(disease_umls['UMLS']
Disease_umls.head(1)
