==> Drug_Repositioning/8_new_training/7_3_RF_independent_1_by_1.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import StratifiedKFold
from sklearn.metrics import confusion_matrix
from sklearn.metrics import roc_auc_score # AUC score
from sklearn.metrics import average_precision_score # AUPR score
from sklearn.metrics import precision_recall_fscore_support # precision, recall
from imblearn.metrics import sensitivity_specificity_support # sensitivity, specificity
from sklearn.metrics import roc_curve # to draw auc curve
from sklearn.metrics import precision_recall_curve # to draw aupr curve
from imblearn.under_sampling import RandomUnderSampler
from imblearn.under_sampling import OneSidedSelection
from imblearn.under_sampling import NearMiss
from itertools import cycle
import matplotlib.pyplot as plt
%matplotlib inline

#2017.3.27. PREDICT independent dataset
indep = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/2_2_indep_descriptor.tsv")
indep_x = indep.values[:, 3:].astype(float)
indep_y = indep.values[:,2].astype(int)

#2017.3.30. PREDICT TN_local + noise dataset
df = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/3_3_2_tn_local_noise_descriptor_predict.tsv")
x_tnNoise = df[df.columns.values[3:].tolist()].values
y_tnNoise = df["association"].values

indep_new_drug.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/PREDICT_data/3_3_indep_new_drug.tsv", sep='\t', index=False)

indep_new_drug_asso[['drug_id', 'disease_id', 'association']].to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/PREDICT_data/3_5_indep_new_drug_asso.tsv", sep='\t', index=False)

#2017.4.3. PREDICT no TN_local
df = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/3_3_tn_local_noise_descriptor_predict.tsv")
x_noTN = df[df.columns.values[3:].tolist()].values
y_noTN = df["association"].values

# 2017.4.3.TN_local + noise 로 train한 모델로 indepndent 예측하는 양상 보기
clf_tnNoise = RandomForestClassifier(n_estimators=500, n_jobs=-1, class_weight='balanced')
clf_tnNoise.fit(x_tnNoise, y_tnNoise)

# 2017.4.3.noTN으로 train한 모델로 indepndent 예측하는 양상 보기
clf_noTN = RandomForestClassifier(n_estimators=500, n_jobs=-1, class_weight='balanced')
clf_noTN.fit(x_noTN, y_noTN)

indep.head()

print clf_tnNoise.predict_proba(indep_x[0].reshape(1,-1))
print clf_noTN.predict_proba(indep_x[0].reshape(1,-1))

print len(indep)

df_result =pd.DataFrame()
cnt = 0
print datetime.now().strftime('%Y-%m-%d %H:%M:%S')
for index, row in indep.iterrows():
    drug_id = row['drug_id']
    disease_id = row['disease_id']
    y = row['association']
    x = row.values[3:].reshape(1,-1)
   
    tnNoise_proba = clf_tnNoise.predict_proba(x)
    tnNoise_label = clf_tnNoise.predict(x)
    noTN_proba =
    clf_noTN.predict_proba(x)
    noTN_label = clf_noTN.predict(x)
    temp_df = pd.DataFrame(data = [[drug_id, disease_id, y, tnNoise_label, tnNoise_proba[:,1][0], noTN_label, noTN_proba[:,1][0]]],
                           columns=['drug_id', 'disease_id', 'y', 'tnNoise_label', 'tnNoise_proba', 'noTN_label','noTN_proba'])
    df_result = df_result.append(temp_df)
    cnt = cnt + 1
    if cnt == 0:
        print "cnt : {}, ".format(cnt) + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
print datetime.now().strftime('%Y-%m-%d %H:%M:%S')

temp = list()
for i in df_result['tnNoise_label']:
    temp.append(i[0])

temp2 = list()
for i in df_result['noTN_label']:
    temp2.append(i[0])

df_temp = df_result

df_temp['tnNoise_label'] = temp
df_temp['noTN_label'] = temp2

df_temp.index = range(0, len(df_temp))
df_temp.head(2)

df_temp.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/8_new_training/PREDICT_data/noTn_vs_noise/independent_1_by_1.tsv", sep='\t', index=False)

# fp
fp_tnNoise = df_temp[(df_temp.y != df_temp.tnNoise_label) & (df_temp.tnNoise_label == 1)]
fp_noTN = df_temp[(df_temp.y != df_temp.noTN_label) & (df_temp.noTN_label == 1)]

# fn
fn_tnNoise = df_temp[(df_temp.y != df_temp.tnNoise_label) & (df_temp.tnNoise_label == 0)]
fn_noTN = df_temp[(df_temp.y != df_temp.noTN_label) & (df_temp.noTN_label == 0)]

# tn
tn_tnNoise = df_temp[(df_temp.y == df_temp.tnNoise_label) & (df_temp.tnNoise_label == 0)]
tn_noTN = df_temp[(df_temp.y == df_temp.noTN_label) & (df_temp.noTN_label == 0)]

# tp
tp_tnNoise = df_temp[(df_temp.y == df_temp.tnNoise_label) & (df_temp.tnNoise_label == 1)]
tp_noTN = df_temp[(df_temp.y == df_temp.noTN_label) & (df_temp.noTN_label == 1)]

path = "/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/8_new_training/PREDICT_data/"
fp_tnNoise.to_csv(path + "noTn_vs_noise/fp_tnNoise.tsv", sep='\t')
fp_noTN.to_csv(path + "noTn_vs_noise/fp_noTN.tsv", sep='\t')
fn_tnNoise.to_csv(path + "noTn_vs_noise/fn_tnNoise.tsv", sep='\t')
fn_noTN.to_csv(path + "noTn_vs_noise/fn_noTN.tsv", sep='\t')
tn_tnNoise.to_csv(path + "noTn_vs_noise/tn_tnNoise.tsv", sep='\t')
tn_noTN.to_csv(path + "noTn_vs_noise/tn_noTN.tsv", sep='\t')
tp_tnNoise.to_csv(path + "noTn_vs_noise/tp_tnNoise.tsv", sep='\t')
tp_noTN.to_csv(path + "noTn_vs_noise/tp_noTN.tsv", sep='\t')

print 'fp : {}, {}'.format(len(fp_noTN), len(fp_tnNoise))
print 'fn : {}, {}'.format(len(fn_noTN), len(fn_tnNoise))
print 'tn : {}, {}'.format(len(tn_noTN), len(tn_tnNoise))
print 'tp : {}, {}'.format(len(tp_noTN), len(tp_tnNoise))

# no TN만 Fn인거
fn_noTN[~fn_noTN.index.isin(fn_tnNoise.index)]

# no TN만 Fp인거
fp_noTN[~fp_noTN.index.isin(fp_tnNoise.index)]

# noise만 FP인거
fp_tnNoise[~fp_tnNoise.index.isin(fp_noTN.index)]

# 공통 fp
union_fp = fp_tnNoise[fp_tnNoise.index.isin(fp_noTN.index)]
union_fp

union_fp.to_csv(path + "noTn_vs_noise/union_fp.tsv", sep='\t')

indep[indep.index.isin(union_fp.index)].to_csv(path + "noTn_vs_noise/union_fp_descriptor.tsv", sep='\t')

tp_noTN

# 공통 fn
fn_noTN[fn_noTN.index.isin(fn_tnNoise.index)]

def Random_Forest_independent_graph(X, y, indep_X, indep_y, user_estimator):

    folds_results = dict()
    draw_results = {'fpr':[], 'tpr':[], 'precision_vec':[], 'recall_vec':[]}
    fp_results = dict()
    
    classifier = RandomForestClassifier(n_estimators=user_estimator, n_jobs=-1, class_weight='balanced')
    classifier.fit(X, y)
    indep_y_predicted_proba = classifier.predict_proba(indep_X)
    indep_y_predicted_label = classifier.predict(indep_X)
    fp_results['predicted_proba'] = indep_y_predicted_proba
    fp_results['Predicted_label'] = indep_y_predicted_label
    fp_results['y_true'] = indep_y

    # Accuracy
    indep_acc = classifier.score(indep_X, indep_y)
    folds_results['acc']=indep_acc
    
    # AUC
    indep_auc = roc_auc_score(indep_y, indep_y_predicted_proba[:,1])
    folds_results['auc']=indep_auc
    
    # Sensitivity, Specificity
    indep_sn, indep_sp, support = sensitivity_specificity_support(indep_y, indep_y_predicted_label)
    folds_results['sn']=indep_sn
    folds_results['sp']=indep_sp
    
    # AUPR
    indep_aupr = average_precision_score(indep_y, indep_y_predicted_proba[:,1])
    folds_results['aupr']=indep_aupr
    
    # Precision, Recall
    indep_precision, indep_recall, _, _ = precision_recall_fscore_support(indep_y, indep_y_predicted_label, average = 'binary')
    folds_results['precision']=indep_precision
    folds_results['recall']=indep_recall
    
    # Confusion Matirx
    indep_confusion_matirx = confusion_matrix(indep_y, indep_y_predicted_label)
    folds_results['confusion_matrix']=indep_confusion_matirx
    
    # draw graph
    indep_fpr, indep_tpr, thresholds = roc_curve(indep_y, indep_y_predicted_proba[:, 1], pos_label=1)
    draw_results['fpr'].append(indep_fpr)
    draw_results['tpr'].append(indep_tpr)
        
    indep_precision_vec, indep_recall_vec, _ = precision_recall_curve(indep_y, indep_y_predicted_proba[:, 1])
    draw_results['precision_vec'].append(indep_precision_vec)
    draw_results['recall_vec'].append(indep_recall_vec)

    return folds_results, draw_results, fp_results

# Random_Forest_independent_graph + 파일 입출력 + 1:1
# PREDICT dataset local no tn indep , indep1 10-fold, 1:1  2017.4.3. 작성
print datetime.now().strftime('%Y-%m-%d %H:%M:%S')
path = '/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/8_new_training/PREDICT_data/'
with open(path + "2_9_RF_no_tn_indep.txt", 'w') as fd :
    user_estimator = 500
    
    fd.write("<independent>\r\n")
    folds_results, draw_results, fp_results = Random_Forest_independent_graph(x_whole_data, y_whole_data, indep_x, indep_y, user_estimator)
    draw_plot([draw_results], folds_results, user_estimator)
    write_output(fd, folds_results, user_estimator)
    #--------------------------------------------------------------
    fd.write("<1:1 independent>\r\n")
    independent_count = 10
    indep_results = dict()
    indep_draw = list()
    for i in range(independent_count):
        x_resampled, y_resampled = rus.fit_sample(x_whole_data, y_whole_data)
        folds_results2, draw_results2, indep_results2= Random_Forest_independent_graph(x_resampled, y_resampled, indep_x, indep_y, user_estimator)
        indep_results[i] = folds_results2
        indep_draw.append(draw_results2)    
    seperated = seperate_results(indep_results)
    final_result = average_result(seperated)
    draw_plot(indep_draw, final_result, user_estimator)
    write_output(fd, final_result, user_estimator)    
print datetime.now().strftime('%Y-%m-%d %H:%M:%S')
